refactor(dto): Standardize on float for numeric API responses

+++ src/services/query_service/app/dtos/cashflow_dto.py
@@ -1,3 +1,4 @@
+# src/services/query_service/app/dtos/cashflow_dto.py
 from pydantic import BaseModel, Field, ConfigDict
 from decimal import Decimal
 from datetime import date
@@ -7,7 +8,7 @@ class CashflowRecord(BaseModel):
     Represents the cashflow details associated with a transaction
     for API responses.
     """
-    amount: Decimal
+    amount: float
     currency: str
     classification: str
     timing: str
